requirements.txt
annotated-types==0.7.0
anyio==4.10.0
azure-core==1.35.0
azure-functions==1.23.0
azure-storage-blob==12.26.0
cachetools==6.2.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
distro==1.9.0
fastapi==0.116.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
jiter==0.10.0
lxml==6.0.0
MarkupSafe==3.0.2
msal==1.33.0
openai==1.99.9
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
PyPDF2==3.0.1
pypdfium2==4.30.0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
requests==2.32.4
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
Werkzeug==3.1.3
